# skattjakt/apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: Skattjakt API
  version: 0.1.0
  summary: Tax Recovery & Opportunity Engine for Swedish limited companies
  description: |
    Skattjakt analyses a Swedish limited company's year-end material and returns
    a structured list of things worth investigating — potential tax positions,
    deductions, accruals, misclassifications and control points.

    This contract is the source of truth for the HTTP surface (section 17).

    **What Skattjakt is not.** It is not a bookkeeping system, a filing system,
    an audit, or legal advice. It issues no binding tax opinion and guarantees
    no saving. Every response carries a `disclaimer` field saying so, and
    clients must display it.

    **How to read a finding.** An `opportunity` is a question worth asking, not
    a conclusion. Its `status` says how far the system got:

      - `identified`   — strong evidence and a reviewed rule.
      - `investigate`  — a potential opportunity, information missing.
      - `verify`       — the rule or calculation needs checking. Everything
                         produced by the rule set shipped today lands here or
                         lower, because that rule set has not been reviewed by
                         a qualified adviser.
      - `warning`      — something to look at, whether or not it saves money.
      - `rejected`     — did not survive falsification; retained for audit.

...
    öre. There is no single-figure field, by design.

  contact:
...
servers:
...
paths:
...
  /v1/documents:
    get:
      parameters:
        - name: limit
          in: query
          required: false
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
          description: Clamped rather than refused if it is out of range.
        - name: cursor
          in: query
          required: false
          schema: { type: string }
          description: |
            Opaque. Pass `next_cursor` from the previous page; do not construct
            one. Keyset rather than offset, so a document uploaded while you are
            paging does not shift or duplicate later rows.
      tags: [Documents]
      summary: List stored document versions
      responses:
        "200": { description: The company's documents. }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "501": { $ref: "#/components/responses/NoPersistence" }
    post:
      tags: [Documents]
      summary: Upload a document
      description: |
        Stores the bytes immutably and records version 1. The document is parsed
        before it is stored, so a file that cannot be read is refused rather than
        kept as an unusable blob. The storage key is derived from the tenant, the
        document and the content hash — never from the filename.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              allOf:
                - $ref: "#/components/schemas/DocumentUpload"
                - type: object
                  properties:
                    kind:
                      type: string
                      enum: [annual_accounts, income_statement, balance_sheet, general_ledger, tax_return]
                    accounts_state: { $ref: "#/components/schemas/AccountsState" }
      responses:
        "201":
          description: Stored.
          content:
            application/json:
              schema:
                type: object
                required: [document_version_id, sha256]
                properties:
                  document_id: { type: string, format: uuid }
                  document_version_id: { type: string, format: uuid }
                  sha256: { type: string }
                  byte_size: { type: integer }
                  page_count: { type: [integer, "null"] }
                  unreadable_pages:
                    type: array
                    items: { type: integer }
        "400":
          description: Unreadable, unsupported, or not what its content type claims.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Problem" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "501": { $ref: "#/components/responses/NoPersistence" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        A bearer token scoped to one company. The token determines the tenant;
        a company identifier in the request body never widens that scope.
  responses:
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
...
    NoPersistence:
      description: |
        No database is configured, so this endpoint cannot work. The stateless
        `POST /v1/analyses` with inline documents remains available.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Problem" }
  schemas:
...
    Problem:
      type: object
      description: |
        Every error body, as `application/problem+json`.

        **Branch on `code`.** It is a stable machine-readable string and the set
        of possible values is pinned by a test, so it cannot change without a
        deliberate contract update. `detail` is prose written for a person and
        will be translated; branching on it breaks the day it is.
      required: [code, title, detail, status]
      properties:
        code:
          type: string
          enum:
            - account_temporarily_locked
            - admin_credential_required
            - already_exists
            - ambiguous_document
            - analysis_failed
            - analysis_is_not_finished
            - authentication_unavailable
            - consent_required
            - document_too_large
            - empty_document
            - insufficient_permission
            - internal_error
            - invalid_accounts_state
            - invalid_base64
            - invalid_credentials
            - invalid_cursor
            - invalid_fiscal_year
            - invalid_idempotency_key
            - invalid_organisationsnummer
            - invalid_request
            - invalid_simulation_model
            - no_company
            - no_documents
            - not_a_session
            - not_found
            - not_what_was_bought
            - nothing_was_uploaded
            - order_not_cancellable
            - order_not_deliverable_yet
            - order_not_payable
            - password_rejected
            - payment_provider_unavailable
            - payment_rejected
            - payment_required
            - payments_not_configured
            - persistence_is_not_configured
            - product_not_available
            - provider_required
            - rate_limited
            - simulation_cancelled
            - simulation_cannot_run
            - simulation_too_slow_to_answer_directly
            - storage_failure
            - tax_year_not_covered
            - the_job_queue_is_not_configured
            - the_session_cannot_be_refreshed
            - the_upload_does_not_match_its_ticket
            - too_many_simulations_at_once
            - unauthorized
            - unknown_audience
            - unknown_delivery_choice
            - unknown_document
            - unknown_product
            - unknown_push_provider
            - unknown_role
            - unknown_value
            - unreadable_document
            - unsupported_document_type
            - wrong_credential
        title: { type: string }
        detail:
          type: string
          description: Safe to show a customer.
        status: { type: integer }
...
    DocumentUpload:
      type: object
      required: [filename, mime_type]
      description: Exactly one of `text` or `content_base64` must be supplied.
      properties:
        filename: { type: string, example: bokslut-2025.pdf }
        mime_type:
          type: string
          enum: [application/pdf, text/csv, text/plain, application/sie]
        text:
          type: [string, "null"]
          description: UTF-8 content, for text formats.
        content_base64:
          type: [string, "null"]
          description: Base64-encoded bytes, for binary formats such as PDF.
    AccountsState:
      type: string
      enum: [preliminary, final, unknown]
      default: preliminary
      description: |
        Whether the accounts are settled. A preliminary year-end is the primary
        case Skattjakt is built for, and it changes what may be concluded.
...
security:
  - bearerAuth: []